+++ R/tm_g_distribution.R
@@ -60,7 +60,7 @@
 #' interactive <- function() TRUE
 #' {{ next_example }}
 # nolint start: line_length_linter.
-#' @examplesIf require("ggpmisc", quietly = TRUE) && require("ggpp", quietly = TRUE) && require("goftest", quietly = TRUE) && require("MASS", quietly = TRUE) && require("broom", quietly = TRUE)
+#' @examples
 # nolint end: line_length_linter.
 #' # general data example
 #' data <- teal_data()
@@ -88,7 +88,7 @@
 #' interactive <- function() TRUE
 #' {{ next_example }}
 # nolint start: line_length_linter.
-#' @examplesIf require("ggpmisc", quietly = TRUE) && require("ggpp", quietly = TRUE) && require("goftest", quietly = TRUE) && require("MASS", quietly = TRUE) && require("broom", quietly = TRUE)
+#' @examples
 # nolint end: line_length_linter.
 #' # CDISC data example
 #' data <- teal_data()
@@ -150,18 +150,9 @@ tm_g_distribution <- function(label = "Distribution Module",
                               plot_width = NULL,
                               pre_output = NULL,
                               post_output = NULL,
-                              decorators = NULL) {
+                              transformators = list(),
+                              decorators = list()) {
   message("Initializing tm_g_distribution")
-
-  # Requires Suggested packages
-  extra_packages <- c("ggpmisc", "ggpp", "goftest", "MASS", "broom")
-  missing_packages <- Filter(function(x) !requireNamespace(x, quietly = TRUE), extra_packages)
-  if (length(missing_packages) > 0L) {
-    stop(sprintf(
-      "Cannot load package(s): %s.\nInstall or restart your session.",
-      toString(missing_packages)
-    ))
-  }
 
   # Normalize the parameters
   if (inherits(dist_var, "data_extract_spec")) dist_var <- list(dist_var)
@@ -204,7 +195,7 @@ tm_g_distribution <- function(label = "Distribution Module",
 
   available_decorators <- c("histogram_plot", "qq_plot", "test_table", "summary_table")
   decorators <- normalize_decorators(decorators)
-  assert_decorators(decorators, null.ok = TRUE, names = available_decorators)
+  assert_decorators(decorators, names = available_decorators)
 
   # End of assertions
 
@@ -231,6 +222,7 @@ tm_g_distribution <- function(label = "Distribution Module",
     ),
     ui = ui_distribution,
     ui_args = args,
+    transformators = transformators,
     datanames = teal.transform::get_extract_datanames(data_extract_list)
   )
   attr(ans, "teal_bookmarkable") <- TRUE
@@ -1227,7 +1219,7 @@ srv_distribution <- function(id,
               expr = {
                 test_table_data <- ANL %>%
                   dplyr::select(dist_var) %>%
-                  with(., broom::glance(do.call(test, args))) %>%
+                  with(., generics::glance(do.call(test, args))) %>%
                   dplyr::mutate_if(is.numeric, round, 3)
               },
               env = env
@@ -1241,7 +1233,7 @@ srv_distribution <- function(id,
                 test_table_data <- ANL %>%
                   dplyr::select(dist_var, s_var, g_var) %>%
                   dplyr::group_by_at(dplyr::vars(dplyr::any_of(groups))) %>%
-                  dplyr::do(tests = broom::glance(do.call(test, args))) %>%
+                  dplyr::do(tests = generics::glance(do.call(test, args))) %>%
                   tidyr::unnest(tests) %>%
                   dplyr::mutate_if(is.numeric, round, 3)
               },