+++ R/tm_t_crosstable.R
@@ -37,7 +37,7 @@
 #' library(teal.modules.general)
 #' interactive <- function() TRUE
 #' {{ next_example }}
-#' @examplesIf require("rtables", quietly = TRUE)
+#' @examples
 #' # general data example
 #' data <- teal_data()
 #' data <- within(data, {
@@ -86,7 +86,7 @@
 #' library(teal.modules.general)
 #' interactive <- function() TRUE
 #' {{ next_example }}
-#' @examplesIf require("rtables", quietly = TRUE)
+#' @examples
 #' # CDISC data example
 #' data <- teal_data()
 #' data <- within(data, {
@@ -143,13 +143,9 @@ tm_t_crosstable <- function(label = "Cross Table",
                             pre_output = NULL,
                             post_output = NULL,
                             basic_table_args = teal.widgets::basic_table_args(),
-                            decorators = NULL) {
+                            transformators = list(),
+                            decorators = list()) {
   message("Initializing tm_t_crosstable")
-
-  # Requires Suggested packages
-  if (!requireNamespace("rtables", quietly = TRUE)) {
-    stop("Cannot load rtables - please install the package or restart your session.")
-  }
 
   # Normalize the parameters
   if (inherits(x, "data_extract_spec")) x <- list(x)
@@ -169,7 +165,7 @@ tm_t_crosstable <- function(label = "Cross Table",
   checkmate::assert_class(basic_table_args, classes = "basic_table_args")
 
   decorators <- normalize_decorators(decorators)
-  assert_decorators(decorators, null.ok = TRUE, "plot")
+  assert_decorators(decorators, "plot")
   # End of assertions
 
   # Make UI args
@@ -189,6 +185,7 @@ tm_t_crosstable <- function(label = "Cross Table",
     ui = ui_t_crosstable,
     ui_args = ui_args,
     server_args = server_args,
+    transformators = transformators,
     datanames = teal.transform::get_extract_datanames(list(x = x, y = y))
   )
   attr(ans, "teal_bookmarkable") <- TRUE