+++ R/utils.R
@@ -34,8 +34,8 @@
 #' - When the length of `size` is three: the plot points size are dynamically adjusted based on
 #' vector of `value`, `min`, and `max`.
 #' @param decorators `r lifecycle::badge("experimental")`
-#' (`list` of `teal_transform_module`, named `list` of `teal_transform_module` or `NULL`) optional,
-#' if not `NULL`, decorator for tables or plots included in the module.
+#' (`list` of `teal_transform_module`, named `list` of `teal_transform_module`) optional,
+#' decorator for tables or plots included in the module output reported.
 #' When a named list of `teal_transform_module`, the decorators are applied to the respective output objects.
 #'
 #' Otherwise, the decorators are applied to all objects, which is equivalent as using the name `default`.
@@ -342,14 +342,9 @@ ui_decorate_teal_data <- function(id, decorators, ...) {
 
 #' Internal function to check if decorators is a valid object
 #' @noRd
-check_decorators <- function(x, names = NULL, null.ok = FALSE) { # nolint: object_name.
-  checkmate::qassert(null.ok, "B1")
+check_decorators <- function(x, names = NULL) { # nolint: object_name.
 
-  check_message <- checkmate::check_list(
-    x,
-    null.ok = null.ok,
-    names = "named"
-  )
+  check_message <- checkmate::check_list(x, names = "named")
 
   if (!is.null(names)) {
     check_message <- if (isTRUE(check_message)) {
@@ -373,7 +368,6 @@ check_decorators <- function(x, names = NULL, null.ok = FALSE) { # nolint: objec
     x,
     checkmate::test_list,
     types = "teal_transform_module",
-    null.ok = TRUE,
     FUN.VALUE = logical(1L)
   )
 
@@ -411,7 +405,7 @@ select_decorators <- function(decorators, scope) {
 #' @return A named list of lists with `teal_transform_module` objects.
 #' @keywords internal
 normalize_decorators <- function(decorators) {
-  if (checkmate::test_list(decorators, "teal_transform_module", null.ok = TRUE)) {
+  if (checkmate::test_list(decorators, "teal_transform_module")) {
     if (checkmate::test_names(names(decorators))) {
       lapply(decorators, list)
     } else {