+++ tests/testthat/test-shinytest2-tm_front_page.R
@@ -14,8 +14,7 @@ app_driver_tm_front_page <- function() {
       ),
       tables = list("MTCARS" = head(mtcars, 5), "IRIS" = head(iris, 5)),
       additional_tags = HTML("Additional HTML or shiny tags go here"),
-      footnotes = "This is a footnote",
-      show_metadata = TRUE
+      footnotes = "This is a footnote"
     ),
     timeout = 3000
   )